fix: 돌림판 최초1회만 로딩하도록 변경
